hookutils: get_installer: fix error handling in macOS-specific checks
Fix error handling in macOS-specific checks in `get_installer`
hook utility function; when trying to run `port` command via
`subprocess`, catch and ignore all instances of `Exception`,
instead of `ExecCommandFailed` (which is a left-over from times
when `exec_command_stdout` was used).

This fixes uncaught `FileNotFoundError: [Errno 2] No such file or
directory: 'port'` error when this part of the code happens to
be reached.