bootloader: attempt to fix console hiding/minimization with Windows Terminal
When trying to hide or minimize console, perform the `ShowWindow`
call in five attempts and with 100 ms delay.

This aims to alleviate timing issues when Windows Terminal is used
as default terminal app; sometimes, we end up calling `ShowWindow`
too soon (apparently before the terminal window is ready for such
requests), and our request to hide/minimize the window ends up
being ignored. On my test system, the operation seems to succeed
on the second attempt, and worst-case, we end up delaying the
program execution for 500 ms.

Sidenote: with contemporary versions of Windows Terminal, only
minimization is supported (hide requests end up only minimizing
the window as well).