tests: have signal-forward/ignore test use a single program build
The refactoring of signal-forward/ignore test in the previous
commit now allows us to build a single instance of the program
per test variant (forward/ignore) and manually run all signal
tests against it.

This does mean that if handling of particular signal fails, we
will need to sift through output of all program runs. But on the
other hand, it saves us from having to build 33 instances of the
same test program in onefile mode.