tests: have frozen stdlib modules test use `importlib.import_module`
Have the frozen stdlib modules test script use `importlib.import_module`
instead of `__import__` to import the frozen modules. This ensures
that `importlib` is always imported, which seems to be necessary
for `_frozen_importlib` to get its `__file__` attribute set.

In python versions < 3.14, importing `pprint` seems to have resulted
in `importlib` being imported, but with 3.14.0a1 this is not the
case anymore, leading to test script failure due to missing
`__file__` attribute on `_frozen_importlib` when running the
test script unfrozen.