depend: exclude cygwin1.dll when building under Cygwin
The `cygwin1.dll` is a rather problematic one.

If we want to run the frozen application *outside of a Cygwin
environment*, we need to collect it. However, because the bootloader
is dynamically linked against it, we need to collect it *next to the
executable*, instead of into the `_internal` directory (onedir builds)
or the `PKG` archive (onefile builds). Otherwise, the executable will
still fail to launch with missing-DLL-error.

On the other hand, if we do collect the DLL (next to the executable),
the application crashes when it is launched *inside a Cygwin environment*
(at least with cygwin 3.5.5); presumably because the duplicated instances
of the DLL conflict with each other. The Cygwin FAQ seems to confirm
this, by stating that each instance of the DLL constitutes its own
Cygwin installation/environment, and that multiple instances are allowed
only as long as they are strictly separated.

Since we cannot guess what the user's end-goal is, we should prevent
`cygwin1.dll` from being collected, and write a new documentation section
for Cygwin under "Platform-specific Notes" to tell users that they should
manually copy the DLL next to the executable if they want the
application to run outside of an existing Cygwin environment.