tests: remove unused scripts/pyi_lib_PyQt5-uic.py
Remove `scripts/pyi_lib_PyQt5-uic.py` script, which is not used by
any test. The data file it refers to also non-existent.

The test this script belonged to was likely extended into what is
now `test_Qt_Ui_file` (which covers all Qt bindings packages), and
at that point, the script file was not removed.