conftest: automatically skip interactive test if psutil is unavailable
When `psutil` is unavailable, have the `AppBuilder.test_script()`
automatically skip the test if `runtime` argument is specified
(i.e., the test program is interactive and we are supposed to
terminate it after the specified amount of time).

The process tree cleanup can be properly implemented only with
`psutil`, so to avoid resource leaks (both processes and temporary
application directories of onefile builds), skip the test. This
should affect only the platforms where `psutil` is unavailable,
such as `cygwin`.