conftest: do not redirect stdout/stderr in AppBuilder._run_executable
Remove the redirection of stdout/stderr to pipes when spawning
subprocess with test frozen application in `AppBuilder._run_executable`,
and instead let the subprocess use the parents stream.

This means that if we run the tests without pytest's capture, we
can see the test program's output "in real-time" rather than after
the program finishes. It also means that we can see the test's
output even if we run into problems with test cleanup and fail to
obtain the streams' data via `Popen.communicate()` call (for example,
because end-of-file is never received due to a rogue child process
of the application process).

On the other hand, this redirect meant that leaked (grand)child
processes could not affect the global test stdout/stderr. If this
proves to be a problem, we can re-introduce the old behavior, but
add an environment-variable-based switch to turn it off for debugging
purposes. It is worth noting, however, that some of our tests also
re-run the built executables on their own, and usually do not
redirect the streams...