ci: cygwin: install up-to-date sphinx from PyPI
For some reason, the Cygwin `python39-pip` package depends on
`python39-sphinx`, which is an old `sphinx` v4.4.0 package. This
version contains a bug caused  by `open()` without specified
`encoding`, which was fixed in sphinx >= v5.0.

This bug is causing our `sphinx` test program to fail (even if ran
unfrozen) with

```
Encoding error:
'ascii' codec can't decode byte 0xc3 in position 213: ordinal not in range(128)
```

due to `locale.getpreferredencoding()` in our Cygwin CI being `ANSI_X3.4-1968`.

To fix this, install the up-to-date `sphinx` from PyPI (as an
alternative to uninstalling `sphinx` or explicitly skipping the
said `sphinx` test when running under Cygwin).