bootloader: Cygwin: onedir application does not need to restart itself
Under Cygwin, we can set the DLL search path without having to
restart the process, so avoid doing that.

This is because the `LD_LIBRARY_PATH` is used by `dlopen()` when
trying to resolve library that was given only by a basename (without
path), while the linked dependencies of binaries and loaded DLLs
are resolved by Windows loader, whose search path is controlled
by `SetDllDirectoryW()`.

Therefore, we need to apply search path modifications using both
mechanisms, but can (in contrast to `LD_LIBRARY_PATH` in general
POSIX codepath) avoid restating the process.

Failing to call `SetDllDirectoryW` results in dependencies
of python shared library not being resolved when running the
frozen application outside of the Cygwin environment.

Failing to set `LD_LIBRARY_PATH` seems to cause segmentation
faults in worker processes when `multiprocessing` is used
(both inside and outside of the Cygwin environment).