conftest: remove debug message with process ID after starting executable
In `AppBuilder._run_executable`, remove the `_display_mesage` call
that prints process ID after the test program executable has been
launched.

This message was added when we were still redirect this subprocess'
outputs to pipes; now that we have stopped doing so, the additional
printed message might interfere with tests that expect to capture
specific (multi-line) messages from bootloader and/or test application.
Therefore, refrain from having the test fixture print additional
messages from the moment the test program is started up to the
point where it exits on its own or we have to kill it due to timeout.
Plus, the process ID can be easily inferred from the bootloader messages.