building: rework the construction of base_library.zip
Rework the way `base_library.zip` is constructed. Instead of constructing
it at the start of the `Analysis.assemble()` and then filtering the
corresponding modules out of the pure-PYMODULE TOC, we propagate those
modules through the main python-module processing part of the
`Analysis.assemble()`, let it decide whether a module should end up
in the PYZ or the `base_library.zip` archive, and construct the zip archive
afterwards.

This approach has several benefits.

It allows us to apply the `module_collection_mode` setting to modules
that end up in the `base_library.zip`; i.e., collect the source .py file
to the filesystem in addition to .pyc collected into the archive, or
collect just the source .py file, or collect just the .pyc file outside
of the archive. The main motivation is ability to collect source .py
files, which might be needed by frameworks that attempt to perform
aggressive recursive introspection.

It allows us to consistently apply optimization flags for .pyc modules,
regardless of whether they end up in the PYZ or `base_library.zip` archive.

When building with `noarchive=True` option, we now construct neither
the PYZ nor `base_library.zip` - all modules are collected as individual
.pyc files.

We do not select modules for `base_library.zip` via regex constructed
from `PY3_BASE_MODULES` anymore; instead, we construct a set for name
lookup based on all nodes obtained during analysis of base-library
modules, and use this set to decide into which archive to put a
module. This allows us to transparently capture dependencies of
base-library modules that might not be listed in `PY3_BASE_MODULES`
themselves, and ensure that they also end up in `base_library.zip`.