building: warn if module cannot be collected as source .py file
When `Analysis.assemble` processes `PYMODULE` entry that is supposed to
be collected as a source .py file (in addition / instead of PYZ), have
it check the module's source path. If it has a .pyc suffix, then the
module's source is unavailable and we cannot collect it - emit a warning
and skip the entry.

Previously, we would either end up collect the .pyc file if it existed
as a file, or emitted a warning during `PKG`/`COLLECT` target assembly
if it did not (for example, for .pyc modules in the `python3x.zip`
archive when using embeddable Windows python build).