ci: replace direct pip invocations with python -m pip
Replace all remaining direct `pip` invocations with `python -m pip`.
This is done both for the sake of consistency, and to make it easier
to adapt the workflow steps in new workflows, where `pip` might
end up pointing to a different python installation (either because
we forgot to install it in our target environment, or because the
target environment does not create the shim).