hookutils: have Tcl/Tk helper gracefully handle built-in _tkinter module
Have our `TclTkInfo` hook utility class gracefully handle cases
when `_tkinter` is a built-in instead of an extension module, and
thus does not have a `__file__` attribute.

Most notable example of this are `indygreg`'s `python-build-standalone`
CPython builds for macOS and Linux, where python seems to be statically
linked against Tcl and Tk (and other) libraries.

Therefore, if we fail to obtain `__file__` attribute of the `_tkinter`
module, skip attempts at determining the location of Tcl and Tk shared
libraries, as those are likely not applicable. This allows us to
initialize other member variables of `TclTkInfo` structure, and properly
collect `tkinter` and its associated Tcl/Tk resources even when
`_tkinter` module is a built-in.

In the splash screen building codepath, add a check for valid Tcl and Tk
shared library paths, as those are hard requirements for the splash screen
functionality. Raise a descriptive error if either of the libraries
could not be resolved. In cases when path to `_tkinter` extension module
is also unavailable (due to it being a built-in), extend the error message
with the note that shared Tcl/Tk libraries are required, while the
python build was likely statically linked against Tcl/Tk, and is thus
inherently incompatible with splash screen.