hookutils: gi: improve error handling in GI module info retrieval
Instead of effectively suppressing error/warning messages by
turning them into debug messages, display all exception messages
from the isolated `_get_module_info()` function call as warnings,
except for the error raised when the module itself (or the
requested version of the module) is not available.

This way, we do not spam warnings when running pre-safe-import hooks
for potentially unavailable modules, but we display warnings when
for example `GIRepository` cannot be imported because of the missing
typelib.