building: ensure macOS executables have unique Mach-O image UUIDs
When assembling the executable for macOS, overwrite the UUIDs stored in
LC_UUID commands in all architecture slices with custom values that
are derived from the original (bootloader executable's) UIID and the
hash of the embedded PKG archive.

This ensures that the Mach-O image UUIDs are unique for different
application executables, as per Apple's TN3178, while also keeping
them reproducible (as long as bootloader executable and the PKG
archive contents remain unchanged).

Valid and unique UUIDs seem to be mandated by certain features on
contemporary macOS versions, for example, local network privacy
that was introduced in macOS 15 (see TN3179).