building: PKG: override the dest name of PYZ archive with PYZ.pyz
When passing the PYZ archive into PKG writer's TOC, normalize the
destination name into PYZ.pyz instead of using the original basename.

The bootloader searches for PYZ using the typecode and implicitly
expects that only one entry is present, so the name should not
really matter.

Having a fixed destination name allows us to create reproducible
repeated builds within the same process (for example, as part of
the tests on our CI), as the default PYZ names include a number
that is automatically incremented across subsequent builds.