hookutils: add get_installer_for_dist as replacement for get_installer
The problem with `get_installer` is that it takes an importable
module/package name as an argument, and must therefore map that
to dist name before trying to read `INSTALLER` file from the dist's
metadata.

The mapping is performed via top-level importable names, which are
listed in `top_level.txt` file in the metadata; however, this file
is not mandatory, and if not present, we cannot map the imprortable
name to the dist. Most notably, this is the case with contemporary
versions of `numpy` (>= 2.0.0). This affects only python 3.10 - in
later versions, stdlib `importlib.metadata` added support for
inferring top-level importable names; in earlier versions, we use
`importlib-metadata`, whose contemporary versions (>= 4.8.1) also
provide support for inferring top-level importable names.

To avoid issues with importable-to-dist mapping altogether, we
deprecate `get_installer` and introduce its replacement,
`get_installer_for_dist`, which takes a distribution name and
attempts to read `INSTALLER` file from its metadata.

While the old helper had some additional provisions for detecting
macports and Homebrew based on the module/package location, that
was actually of little use to us.