modulegraph: try to look for .py or .pyi files that accompany extensions
When modulegraph encounters an extension module, have it check
for adjacent `.py` or `.pyi` files (in that order of priority).
If such accompanying source (or interface) file is available, compile
it into code object to perform import analysis on it.

This attempts to improve import analysis of extension modules,
which on their own cannot be analyzed by modulegraph. However,
sometimes cythonized packages ship both binary extensions and
original .py files, or the extensions have accompanying .pyi files;
both of those can be analyzed for imports as if they were regular
source files, and might provide information about at least some,
if not all, imports that are made within the extension.