tests: xfail test_reproducible_subsequent_builds for python < 3.11 on linux
On GHA ubuntu runner, we seem to be getting occasional failures
of `test_reproducible_subsequent_builds` due to differences in
collected modules' bytecode. This happens in spite of the test
first performing a throw-away "calibration" build.

So far, the failures have been observed only on linux runner with
python < 3.11; most often, it fails with 3.9, but 3.8 and 3.10
have also been observed to fail.

Since the lack of .pyc bytecode reproducibility seems to be out of
our control, xfail the test.