bootloader: waf: remove duplicated check for strip
Remove the duplicated check for `strip` utility, which resulted in
duplicated `Checking for program 'strip': ...` message in the build
log.

The first check was done in `wscript` and checks the parent directory
of the C compiler executable as well as locations in PATH. The second
check was performed immediately afterwards, as the `tools/strip.py`
feature/tool was loaded. Because both use the same variable name
(i.e., `STRIP`), the second check was essentially no-op. We now
perform the check for `strip` utility only in the `configure` step
of `tools/strip.py` (where we check both the parent directory of
C compiler executable and locations in PATH).