building: on AIX, pass -X32_64 flag to strip utility
When processing collected binaries on AIX with strip enabled, pass
the `-X32_64` flag to the `strip` utility to have it transparently
process both 32-bit and 64-bit binaries. On one hand, this alleviates
the need for user to set `OBJECT_MODE` environment variable prior
to running PyInstaller (to force 32-bit or 64-bit mode), but on the
other, it also accommodates potential mixed-mode scenarios; for
example, a 32-bit helper program being collected into otherwise 64-bit
application bundle.

In documentation, remove the note that says that OBJECT_MODE needs
to be set when using PyInstaller on AIX (as opposed to building the
bootloader), as it should not be needed there.