bootloader: on POSIX, have parent of onefile with splash restart itself
On POSIX systems where we set library search path by modifying the
`LD_LIBRARY_PATH` (or equivalent) environment variable, have the parent
process of onefile application with splash screen restart itself after
creating the application's temporary directory and setting the environment
variable.

This ensures that the library search modification is in effect in the
parent process when it tries to load Tcl and Tk shared libraries for
the splash screen functionality. Failing to do so results in dependencies
of Tcl and Tk shared libraries being loaded from the system (instead of
the bundled copies), which might result in errors when the system does
not provide the said dependencies. Furthermore, loading Tcl and Tk
dependencies is inconsistent with behavior in onedir codepath, where
library search path modification is also already in effect (due to
process restart) at the point when splash screen is started.