tests: multiprocessing: remove timeouts in AsyncResult.get() calls
Remove 1-second timeout from `multiprocessing.pool.AsyncResult.get()`
calls in test scripts used by `test_multiprocessing_pool` and
`test_multiprocessing_subprocess_environment`.

The timeout should be unnecessary in the first place (we have a
timeout also set on the executable's run-time, as well as the
test itself); it is also too strict for the cases when `pytest`
is ran with number of workers that match or exceed the number
of the CPUs, especially because the tests themselves also spawn
four additional worker processes, easily leading to CPU starvation
and stalling on our CI.

The issue surfaced when running tests with free-theaded python 3.13;
presumably because multiple threads can now further contribute to
the CPU exhaustion and starvation.