tests: run a dummy subprocess at the start of test_pipe_leakage
On Windows, the very first `subprocess.Popen()` seems to open two
additional handles, which seem to remain open for the duration of
the python process.

Therefore, if a subprocess is spawned before `test_pipe_leakage` is
ran (either as a part of a preceding test, as a part of test collection,
or even as a side effect of some 3rd party package having been loaded),
those two extra handles will already be part of the initially-opened
handles. If not, they will show up during this test and skew the results.

Therefore, run a dummy subprocess here to normalize the conditions,
and ensure that the test can be ran on its own.