loader: set module's file path to source .py file if available
Have our `PyiFrozenLoader` set the module's file path (i.e., the
loader's `path` attribute and the module's `__file__` attribute)
to the source .py file if the latter is available (i.e., was
explicitly collected into the frozen application in addition to
the byte-compiled version in the PYZ archive).

This brings our loader's behavior in compliance with the description
of `importlib.abc.ExecutionLoader.get_filename`, which states that,
"If source code is available, then the method should return the path
to the source file, regardless of whether a bytecode was used to load
the module."

Returning the path to the source .py file (instead of fictional
.pyc file in the same location) improves compatibility with 3rd-party
code that naively assumes that the module's `__file__` attribute
always points to the source .py file, and either tries to load it
as a source file, or attempts to perform filesystem-based operations
under assumption that the file and/or its parent directory exist.