rthooks: update `inspect` run-time hook due to changes in PyiFrozenLoader behavior
Update the `inspect` run-time hook to account for the changes made
in the `PyiFrozenLoader` behavior regarding the module file path;
since the loader now sets module file path to source .py files
when they are available, our override of `inspect.getsourcefile()`
needs to do the same.

This applies only to modules that are part of the PYZ archive and
are handled by `PyiFrozenLoader` (but not to the ones that are part
of the `base_library.zip` and are covered by python's own `zipimporter`).