loader: have __file__ attribute use .py suffix instead of .pyc one
Have our `PyiFrozenLoader` use the .py suffix instead of .pyc when
computing the module's file path (i.e., when setting the loader's `path`
attribute and the module's `__file__` attribute).

Historically, PyInstaller's loader/importer has been setting `__file__`
to path where the .pyc file would be located if it were stored as a
file (rather than being collected in the PYZ archive); this way, the
`__file__` attribute can be used in both unfrozen and frozen code to
reference data files and other resources that are adjacent to .py
files in the original, unfrozen packages (and collected into frozen
application using same directory layout). For this purpose, the suffix
of the module file is irrelevant, as typically the parent directory
of the `__file__` attribute is used.

On the other hand, a non-negligible amount of libraries naively assume
that the `__file__` attribute points to the source .py file, and try
to either load it as a source file, or perform some access operation
on it (for example, check for its existence, read creation time, etc.).
Often, the same libraries already require us to collect source .py files
anyway. So having the `__file__` attribute point to (potentially existing)
.py file rather than (non-existant) .pyc file allows us to better
accommodate these libraries by just collecting the source .py file (which
is often required anyway) and without requiring us to forego the byte-compiled
version in the PYZ archive.

Adjust the `inspect` run-time hook for the modified behavior (for
modules handled by `PyiFrozenLoader`, it should also resolve the
relative file paths into .py filenames rather than .pyc ones).

Update the documentation in the couple of places that mention
the `__file__` attribute in the context of frozen application.