hooks: remove hook-packaging.py
Remove the hook for `packaging`, whose only function was to add
`pkg_resources` to hidden imports.

That hidden import results in `pkg_resources` being pulled into the
frozen application (along with our run-time hook for it) even if no
part of the code actually tries to use it. And with up-to-date
`setuptools`,  this results in *run-time* deprecation warning when
our run-time hook for `pkg_resources` tries to import the package
and set up our own provider for it.