bootloader: add support for narrow-char strings in run-time options collection
Have the `PyiRuntimeOptions` and `pyi_runtime_options_read()`
collect the W-flags and X-flags strings into narrow-char string
arrays (`wflags` and `xflags`) when PEP 741 API is available,
and into wide-char string arrays (`wflags_w` and `xflags_w`)
when not.

This minimizes the amount of conversion that will be needed in
PEP 741 configuration codepath (which works with narrow-char
UTF-8 strings) and in the existing PEP 587 configuration codepath
(which works with wide-char strings).