bootloader: fix bug in error-handling code in pyi_pyconfig.c
In `pyi_pyconfig_set_module_search_paths`, fix the check for
failed element conversion, which is using incorrect comparison,
`ret < -1`; change that to `ret != 0`.

Fix a similar problem in `pyi_pyconfig_set_argv`; but also change
the code to jump to end of the function immediately on error. In
this function, we do not need to initialize all elements, because
the array is allocated (and zero-initialized) via  `calloc`.