tests: add test for subpackage exclusion from top-level package hook
Prepare a test for subpackage exclusion from top-level package hook;
if the hook for `mypackage` excludes `mypackage.optional` (because,
for example, that subpackage is heavy-weight but is not really used
by `mypackage` and its other submodules, except for some codepath
that is not executed during normal use scenarios), the said
sub-package should be fully excluded. However, if there is an external
reference to this sub-package (from user's program or some 3rd party
package they are using), this sub-package should be collected, and
the exclusion rule should not prevent collection of sub-modules from
this sub-package.

This latter part is currently missing, and is affecting `sqlalchemy`
(where our hook excludes `sqlalchemy.testing`), as well as older
versions of `numpy` (where our hook excludes `numpy.f2py`); if the
user's program tries to import these excluded subpackages, they end
up with error because the exclusion rule also blocks collection
submodules.

The test consists of two variants. In the first, only top-level
package is included, and we check that the optional subpackage is
not collected; this ensures that exclusion mechanism works as
intended. In the second, the test program imports the optional
subpackage (which works only if its submodule is also collected);
this ensures that collection works as intended when there is an
external reference.

At the moment, the first variant fails, and the second one suceeds
(contrary to what one would expect based on the above text); this
is because several different import patterns are used, and our
exclusion code currently fails to properly interpet one of the
relative imports, which allows the subpackage to slip in. Once
this is fixed, the first variant should succeed and second one
fail. Then, after we ensure that exclusion rules are properly
suspended in the presence of an external reference, both variants
should succeed.