analysis: module exclusion: fix module name construction for relative imports
Fix the construction of module name that is used for evaluation
of module exclusion rules: when relative imports are used, the
adjustment of the base module should be done *before* we append the
`target_module_partname`, not after.

This fixes the input combination:
- source_module='mypackage.mandatory.a'
- target_module_partname='optional'
- level=2
to evaluate into `mypackage.optional` instead of the former
`mypackage.mandatory`.

And more tellingly, the input combination:
- source_module='mypackage.mandatory.a'
- target_module_partname='optional.b'
- level=2
to evaluate into `mypackage.optional.b` instead of former (nonsense)
`mypackage.mandatory.optional`.