analysis: account for hook-excluded but externally-referenced subpackages
When evaluating a module name against the applicable set of exclusion
rules, waive any rule that also applies to the name of the module's
referrer (i.e., the module that imported the module currently
under evaluation).

This aims to fix the problem that arises when a hook for a top-level
package (for exaple, `sqlalchemy`) excludes a subpackage (for example,
`sqlalchemy.testing`) to prevent its collection in the absence of
any external references (i.e., not coming from `sqlalchemy` itself).
When there are actually external references, the collection of
`sqlalchemy.testing` is allowed, but not of its submodules (for
example, `sqlalchemy.testing.config`), leading to module-not-found
errors when `sqlalchemy.testing` tries to import those submodules.

In the above example, we want to suspend the `sqlalchemy.testing`
exclusion rule (from the top-level `sqlalchemy` hook) when module
name is `sqlalchemy.testing.config` and its referrer is
`sqlalchemy.testing`, which is a sign of a subpackage referring
to its submodules. The fact that `sqlalchemy.testing` is a referrer
tells us that it was previously analyzed, in spite of this particular
exclusion rule (i.e., due to an import chain that did not trigger
this exclusion rule; with exclusion rules being tied to referrers
and their parents).