building: move search for python shared lib to the very start of build/analysis
Move the search for python shared library to the very start of
the build / analysis process, i.e., at the beginning of the
`Analysis.assemble`.

This ensures that we raise an error about static-library python
builds or missing shared library as soon as possible, i.e., before
even initializing modulegraph.