modulegraph: work around the use of pkg_resources in altgraph
Add a work-around for `altgraph`'s use of deprecated and
soon-to-be-removed `pkg_resources`, namely the use of
`pkg_resources.require("altgraph")[0].version` to query its
version.

Instead of waiting for updated version of `altgraph` that uses
`importlib.metadata` (which may never arrive at all) or vendoring
a copy with required change, have our vendored copy of `modulegraph`
provide a dummy module with fallback implementation of `require()`
method that allows `altgraph` to query its version, and then remove
this dummy module once `altgraph` is imported.