Prevent altgraph from importing pkg_resources
pkg_resources is scheduled to be removed in 2025-11-30 but altgraph
still uses it to query its own version. There is an open pull request to
altgraph to fix it (ronaldoussoren/altgraph#25) but it's showing no
signs getting merged.

Monkeypatch a fake pkg_resources into sys.modules when importing
altgraph so that its version query doesn't take out PyInstaller come
2025-11-30.