test: Freeze garbage collect during isolated's pipe leakage test
It may be that the pipe leakage test's flaky behaviour is due to a
garbage collection happening mid-test, causing some unrelated handles
to be closed.

Even if that's not the case though, it makes the test more predictable
in the event that PyInstaller.isolated is mistakenly relying on garbage
collect to close handles.