bindepend: revise the search for Python shared library and its restrictions
Revise the POSIX part of the search for Python shared library and the
restrictions it imposes.

When `Py_ENABLE_SHARED` is set to 0, we now always take the fall-back
codepath with best-guess for the shared library name, instead of limiting
it only to the case when `compat.is_conda` is set. However, in this
fall-back codepath, we now search for the shared library only in the
`sys.base_prefix` or the `lib` directory under it (instead of passing
the guess-based name to the common search codepath).

This way, we do not block Python builds that have the same characteristics
as Anaconda Python (separate shared library and `Py_ENABLE_SHARED` set to 0);
for example, the use case from #9273 that evades our `compat.is_conda`
check due to removal of `conda-meta` directory. But there could be other
Python builds like that in the wild...

On the other hand, we further limit the search scope for such cases;
for example, if we encountered Anaconda Python without shared library
installed (since technically, it is not required to be present), we
should now raise an error instead of searching the common library
search path (which could pick up an unrelated Python shared library
of matching version).