hookutils: Tcl/Tk: skip threaded check for Tcl 9
Tcl 9.0 removed the `--enable-threads` build flag, and is always
built with multi-threaded support. Therefore, the `threaded` member
has been completely removed from the `tcl_platform` array.

Therefore, if we encounter Tcl >= 9, we need to skip the check
for `threaded` member in `tcl_platform`, and instead always
set the flag indicating that multi-threading is supported.