hookutils: Tcl/Tk: tweak the search order of Tcl/Tk shared libraries
Have the `_find_tcl_tk_shared_libraries` helper first check if the
library name contains "tk", followed by check for "tcl", rather than
the other way around.

It is unlikely that "tk" will appear in the name of the Tcl shared
library, while "tcl" could appear in the name of the Tk shared
library. For example, Fedora 43 ships both Tcl/Tk 8.6 and 9.0, and
in the latter, the shared libraries are named `libtcl9.0.so` and
`libtcl9tk9.0.so`.