bootloader: splash: add support for Tcl 9
In Tcl 9, various size-related arguments to functions were upgraded
from 32-bit types (`unsigned` or `int`) to 64-bit types (`size_t`
or `ptrdiff_t`) while the names of the symbols remained the same.

Consequently, we now need two prototypes for affected functions,
and bind the function to correct prototype based on the version
of Tcl that we detect at run-time.