building: splash: delay imports of PIL.Image and tcl_tk hookutil module
Delay imports of `PIL.Image` and `PyInstaller.utils.hooks.tcl_tk`, by
moving them from top-level of `PyInstaller.building.splash` module
into `Splash.__init__()` and `Splash.assemble()`.

This way, we can avoid unnecessarily importing these modules when
building an application that does not splash screen enabled (due
to the way the `Splash` target is made available in the .spec file,
we need to import the `splash` module regardless of whether it is
needed or not...)

As a side-effect, the module-level `splash_requirements` list has
been inlined directly into construction of `Splash.splash_requirements`
set in the `Splash` constructor.