bootloader: have our tclInit override run init.tcl
Have our `tclInit` override load and run `init.tcl` script (and add
the said script to the list of files required by splash screen).

Starting with Tk 9.0, some of the Tk commands are apparently expected
to be auto-resolved and loaded. While we could provide our own
implementation for `unknown` command that resolves the couple of
commands used in the minimal set of Tk .tcl modules strictly
required by the splash screen, in case when full Tcl/Tk is
available (either because onedir application uses `tkinter`, or
because splash screen was built with `full_tk` option), this
would need to cover all commands. Therefore, it is probably better
to use the auto-loading implementation from the `init.tcl`.

While this is not required for Tcl/Tk 8.x, the new codepath is
intentionally used there as well, to increase the exposure and
flush out potential issues.