building: ensure suffix in co_filename of the entry-point's code object
When anonymizing path in `co_filename` of the entry-point script's
code object, ensure that it always has a suffix - by taking the suffix
from the original `co_filename` of code object produced by the
`get_code_object()` helper.

In cases when the entry-point script's source file does not have a
suffix, the `get_code_object()` helper adds a `.py` suffix to the
`co_filename` in an attempt to mitigate issues with `traceback`
module mistakenly trying to load POSIX executables as if they were
source files (if the executable happens to be in the current working
directory).

Fixes regression that was introduced in dfa79a8.